+++ app/src/main/java/com/sekusarisu/mdpings/vpings/presentation/components/MDAppTopBar.kt
@@ -44,11 +44,13 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.PreviewLightDark
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import com.sekusarisu.mdpings.R
 import com.sekusarisu.mdpings.ui.theme.MDPingsTheme
 import com.sekusarisu.mdpings.vpings.domain.ServerOrder
 import com.sekusarisu.mdpings.vpings.domain.ServerSortField
@@ -72,7 +74,11 @@ fun MDAppTopBar(
     val scope = rememberCoroutineScope()
     var isDropDownExpanded by remember { mutableStateOf(false) }
     // ServerSortField -> serverSortFieldString
-    val serverSortFieldString = listOf("ServerID", "Index", "Online")
+    val serverSortFieldString = listOf(
+        stringResource(R.string.sort_field_string_serverid),
+        stringResource(R.string.sort_field_string_index),
+        stringResource(R.string.sort_field_string_online)
+    )
     val selectedServerSortField = appSettingsState.appSettings.serverSortField.ordinal
     val serverOrder = appSettingsState.appSettings.serverOrder.ordinal
     val isExpanded = appSettingsState.appSettings.expandedServerListCard